+++ django_asgi_lifespan/signals.py
@@ -5,16 +5,14 @@
 
 from __future__ import annotations
 
-from typing import Annotated, Final
+from typing import Final
 
-import django
+from django.dispatch import Signal
 
-if django.VERSION > (5, 0):  # pragma: no cover
-    from .dispatcher import PatchedSignal as Signal
-else:
-    from django.dispatch import Signal  # type: ignore
+from .compat import CompatAsyncSignal
 
-__all__ = ["asgi_startup", "asgi_shutdown"]
+__all__ = ["asgi_startup", "asgi_lifespan", "asgi_shutdown"]
 
-asgi_startup: Final[Annotated[Signal, "asgi lifespan startup"]] = Signal()
-asgi_shutdown: Final[Annotated[Signal, "asgi lifespan shutdown"]] = Signal()
+asgi_startup: Final[Signal] = Signal()
+asgi_lifespan: Final[CompatAsyncSignal] = CompatAsyncSignal()
+asgi_shutdown: Final[Signal] = Signal()