+++ django_asgi_lifespan/types.py
@@ -1,0 +1,7 @@
+# -*- encoding: utf-8 -*-
+# ! python3
+
+from typing import Any, AsyncContextManager, MutableMapping, TypeAlias
+
+LifespanManager: TypeAlias = AsyncContextManager[MutableMapping[str, Any]]
+State: TypeAlias = MutableMapping[str, Any]