+++ tests/django_test_application/settings.py
@@ -12,3 +12,22 @@ INSTALLED_APPS = [
 USE_TZ = True
 ROOT_URLCONF = "tests.django_test_application.urls"
 SECRET_KEY = "se3ret"
+MIDDLEWARE = [
+    # ...
+    "django_asgi_lifespan.middleware.AsyncMiddleware",
+    # ...
+]
+LOGGING = {
+    "version": 1,
+    "disable_existing_loggers": False,
+    "handlers": {
+        "console": {
+            "class": "logging.StreamHandler",
+        },
+    },
+    "root": {
+        "handlers": ["console"],
+        "level": "INFO",
+    },
+}
+ALLOWED_HOSTS = ["*"]